src/python-legacy/youtubevideostockDonwloader/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.8.0
yt-dlp>=2024.1.0
